refactor(python/timestamp): move package to timestamp_ext and flatten showcase files
