--- samples/python/extensions/timestamp/tests/__init__.py
@@ -1,1 +1,0 @@
-# Test package for the timestamp extension.
