+++ samples/python/extensions/timestamp/timestamp_ext/core.py
