+++ samples/python/extensions/timestamp/timestamp_ext/py.typed
