+++ samples/python/extensions/timestamp/timestamp_ext/server.py
@@ -1,6 +1,7 @@
 from a2a.server.agent_execution import AgentExecutor, RequestContext
 from a2a.server.events.event_queue import Event, EventQueue
-from core import TimestampExtension
+
+from timestamp_ext.core import TimestampExtension
 
 
 def wrap_executor(executor: AgentExecutor, ext: TimestampExtension) -> AgentExecutor:
